+++ lora/T3S3.h
@@ -35,7 +35,7 @@ extern uint8_t spreadingFactor;
 extern int packetSize;
 
 // Store every SNR of every received packet, to later calculate avg using receiveCounter
-extern unsigned long totalSNR;
+extern double totalSNR;
 
 // --------------------------------------------- //
 //              FUNCTIONALITY CODE               //
